requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
SQLAlchemy[asyncio]>=2.0.40,<3.0.0
asyncpg>=0.30.0,<1.0.0
pydantic-settings>=2.8.0,<3.0.0
bcrypt>=4.1.3,<5.0.0
PyJWT>=2.9.0,<3.0.0
email-validator>=2.2.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
httpx>=0.27.2,<1.0.0
celery[redis]>=5.4.0,<6.0.0
redis>=5.0.0,<6.0.0
sse-starlette>=2.0.0,<3.0.0
